+++ game_patch/multi/server_internal.h
@@ -51,6 +51,7 @@ struct ServerAdditionalConfig
     int anticheat_level = 0;
     bool stats_message_enabled = true;
     bool randomize_rotation = false;
+    bool dynamic_rotation = false;
     std::string welcome_message;
     bool weapon_items_give_full_ammo = false;
     float kill_reward_health = 0.0f;