+++ docs/CHANGELOG.md
@@ -46,6 +46,7 @@ Version 1.9.0 (not released yet)
 - Add `map_rand` command to change to a random level on the rotation
 - Add `vote rand` vote type to initiate a level change to a random level on the rotation
 - Add `$DF Random Level Order` option in dedicated server config
+- Add `$DF Dynamic Rotation` option in dedicated server config (shuffles rotation once complete)
 
 Version 1.8.0 (released 2022-09-17)
 -----------------------------------